# openapi.yaml
openapi: 3.0.3
info:
  title: FX Deals Data Warehouse API
  description: |
    Bloomberg FX Deals Data Warehouse REST API for managing foreign exchange deal transactions.
    
    This API allows you to:
    - Create new FX deals with validation
    - Retrieve FX deals by ID or unique deal identifier
    - List all FX deals with pagination support
    - Prevent duplicate deal submissions
    
    **Technology Stack:**
    - Java 17
    - Jakarta EE 10
    - WildFly 30.0.0.Final
    - MySQL 8.0
    - Hibernate 6.2.13
    
  version: 1.0.0
  contact:
    name: Bloomberg FX Deals Team
    email: support@bloomberg.com
  license:
    name: Proprietary
    url: https://www.bloomberg.com

servers:
  - url: http://localhost:8080/fxdeals
    description: Local development server
  - url: http://localhost:8080/fxdeals
    description: Docker container deployment

tags:
  - name: FX Deals
    description: Operations related to FX deal management
  - name: Health
    description: API health check endpoints

paths:
  /api/fx-deals:
    post:
      tags:
        - FX Deals
      summary: Create a new FX deal
      description: |
        Creates a new foreign exchange deal in the system. The deal must have a unique `dealUniqueId`.
        
        **Validation Rules:**
        - Deal Unique ID: Required, must be unique across all deals
        - From Currency: Required, must be a valid 3-letter ISO code (e.g., USD)
        - To Currency: Required, must be a valid 3-letter ISO code (e.g., EUR)
        - Deal Timestamp: Required, cannot be in the future
        - Deal Amount: Required, must be greater than 0.01, max 15 integer digits and 2 decimal places
        - From and To currencies must be different
        
      operationId: createDeal
      requestBody:
        description: FX Deal data to create
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FxDealRequest'
            examples:
              validDeal:
                summary: Valid FX Deal
                value:
                  dealUniqueId: "DEAL-2025-001"
                  fromCurrencyIso: "USD"
                  toCurrencyIso: "EUR"
                  dealTimestamp: "2025-11-13T10:30:00"
                  dealAmount: 1000000.50
              largeDeal:
                summary: Large Amount Deal
                value:
                  dealUniqueId: "DEAL-2025-002"
                  fromCurrencyIso: "GBP"
                  toCurrencyIso: "JPY"
                  dealTimestamp: "2025-11-13T11:45:00"
                  dealAmount: 999999999999999.99
      responses:
        '201':
          description: Deal created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FxDealResponse'
              examples:
                created:
                  value:
                    id: 1
                    dealUniqueId: "DEAL-2025-001"
                    fromCurrencyIso: "USD"
                    toCurrencyIso: "EUR"
                    dealTimestamp: "2025-11-13T10:30:00"
                    dealAmount: 1000000.50
                    createdAt: "2025-11-13T10:30:15.123"
        '400':
          description: Invalid request data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                missingField:
                  summary: Missing required field
                  value:
                    error: "Validation failed"
                    message: "Deal Unique ID is required"
                    timestamp: "2025-11-13T10:30:00"
                invalidCurrency:
                  summary: Invalid currency format
                  value:
                    error: "Validation failed"
                    message: "From Currency must be a valid 3-letter ISO code (e.g., USD)"
                    timestamp: "2025-11-13T10:30:00"
                sameCurrency:
                  summary: Same from/to currency
                  value:
                    error: "Business rule violation"
                    message: "From and To currencies must be different"
                    timestamp: "2025-11-13T10:30:00"
                negativeAmount:
                  summary: Negative or zero amount
                  value:
                    error: "Validation failed"
                    message: "Deal amount must be greater than zero"
                    timestamp: "2025-11-13T10:30:00"
                futureDate:
                  summary: Future timestamp
                  value:
                    error: "Validation failed"
                    message: "Deal timestamp cannot be in the future"
                    timestamp: "2025-11-13T10:30:00"
        '409':
          description: Deal with this unique ID already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                duplicate:
                  value:
                    error: "Duplicate deal"
                    message: "Deal with unique ID 'DEAL-2025-001' already exists"
                    timestamp: "2025-11-13T10:30:00"
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    get:
      tags:
        - FX Deals
      summary: Get all FX deals
      description: |
        Retrieves a paginated list of all FX deals in the system.
        
        **Pagination:**
        - Default offset: 0
        - Default limit: 100
        - Maximum limit: 1000
        
        **Response Headers:**
        - `X-Total-Count`: Total number of deals in the database
        - `X-Offset`: Current offset used
        - `X-Limit`: Current limit used
        
      operationId: getAllDeals
      parameters:
        - name: offset
          in: query
          description: Starting position for pagination (0-based)
          required: false
          schema:
            type: integer
            minimum: 0
            default: 0
          example: 0
        - name: limit
          in: query
          description: Maximum number of results to return (1-1000)
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 1000
            default: 100
          example: 10
      responses:
        '200':
          description: List of FX deals retrieved successfully
          headers:
            X-Total-Count:
              description: Total number of deals in the database
              schema:
                type: integer
              example: 150
            X-Offset:
              description: Current offset used in the query
              schema:
                type: integer
              example: 0
            X-Limit:
              description: Current limit used in the query
              schema:
                type: integer
              example: 10
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/FxDealResponse'
              examples:
                dealsList:
                  value:
                    - id: 1
                      dealUniqueId: "DEAL-2025-001"
                      fromCurrencyIso: "USD"
                      toCurrencyIso: "EUR"
                      dealTimestamp: "2025-11-13T10:30:00"
                      dealAmount: 1000000.50
                      createdAt: "2025-11-13T10:30:15.123"
                    - id: 2
                      dealUniqueId: "DEAL-2025-002"
                      fromCurrencyIso: "GBP"
                      toCurrencyIso: "JPY"
                      dealTimestamp: "2025-11-13T11:45:00"
                      dealAmount: 500000.75
                      createdAt: "2025-11-13T11:45:20.456"
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /api/fx-deals/{id}:
    get:
      tags:
        - FX Deals
      summary: Get FX deal by database ID
      description: Retrieves a specific FX deal using its database ID
      operationId: getDealById
      parameters:
        - name: id
          in: path
          description: Database ID of the FX deal
          required: true
          schema:
            type: integer
            format: int64
          example: 1
      responses:
        '200':
          description: Deal found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FxDealResponse'
              examples:
                found:
                  value:
                    id: 1
                    dealUniqueId: "DEAL-2025-001"
                    fromCurrencyIso: "USD"
                    toCurrencyIso: "EUR"
                    dealTimestamp: "2025-11-13T10:30:00"
                    dealAmount: 1000000.50
                    createdAt: "2025-11-13T10:30:15.123"
        '404':
          description: Deal not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                notFound:
                  value:
                    error: "Not found"
                    message: "Deal with ID 99999 not found"
                    timestamp: "2025-11-13T10:30:00"
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /api/fx-deals/unique/{dealUniqueId}:
    get:
      tags:
        - FX Deals
      summary: Get FX deal by unique deal ID
      description: Retrieves a specific FX deal using its unique deal identifier
      operationId: getDealByUniqueId
      parameters:
        - name: dealUniqueId
          in: path
          description: Unique identifier of the FX deal
          required: true
          schema:
            type: string
          example: "DEAL-2025-001"
      responses:
        '200':
          description: Deal found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FxDealResponse'
              examples:
                found:
                  value:
                    id: 1
                    dealUniqueId: "DEAL-2025-001"
                    fromCurrencyIso: "USD"
                    toCurrencyIso: "EUR"
                    dealTimestamp: "2025-11-13T10:30:00"
                    dealAmount: 1000000.50
                    createdAt: "2025-11-13T10:30:15.123"
        '404':
          description: Deal not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                notFound:
                  value:
                    error: "Not found"
                    message: "Deal with unique ID 'DEAL-NOT-EXISTS' not found"
                    timestamp: "2025-11-13T10:30:00"
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /api/fx-deals/health:
    get:
      tags:
        - Health
      summary: Health check endpoint
      description: Simple endpoint to verify that the API is running
      operationId: healthCheck
      responses:
        '200':
          description: API is healthy
          content:
            text/plain:
              schema:
                type: string
                example: "FX Deals API is running"

components:
  schemas:
    FxDealRequest:
      type: object
      required:
        - dealUniqueId
        - fromCurrencyIso
        - toCurrencyIso
        - dealTimestamp
        - dealAmount
      properties:
        dealUniqueId:
          type: string
          description: Unique identifier for the deal (must be unique across all deals)
          minLength: 1
          example: "DEAL-2025-001"
        fromCurrencyIso:
          type: string
          description: ISO 4217 currency code for the source currency (3 uppercase letters)
          pattern: '^[A-Z]{3}$'
          example: "USD"
        toCurrencyIso:
          type: string
          description: ISO 4217 currency code for the target currency (3 uppercase letters)
          pattern: '^[A-Z]{3}$'
          example: "EUR"
        dealTimestamp:
          type: string
          format: date-time
          description: Timestamp when the deal occurred (ISO 8601 format, cannot be in the future)
          example: "2025-11-13T10:30:00"
        dealAmount:
          type: number
          format: decimal
          description: Deal amount in the source currency (must be > 0, max 15 integer digits, 2 decimal places)
          minimum: 0.01
          maximum: 999999999999999.99
          example: 1000000.50

    FxDealResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
          description: Database-generated unique ID
          example: 1
        dealUniqueId:
          type: string
          description: Unique identifier for the deal
          example: "DEAL-2025-001"
        fromCurrencyIso:
          type: string
          description: ISO 4217 currency code for the source currency
          example: "USD"
        toCurrencyIso:
          type: string
          description: ISO 4217 currency code for the target currency
          example: "EUR"
        dealTimestamp:
          type: string
          format: date-time
          description: Timestamp when the deal occurred
          example: "2025-11-13T10:30:00"
        dealAmount:
          type: number
          format: decimal
          description: Deal amount in the source currency
          example: 1000000.50
        createdAt:
          type: string
          format: date-time
          description: Timestamp when the deal was created in the system
          example: "2025-11-13T10:30:15.123"

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          description: Error type or category
          example: "Validation failed"
        message:
          type: string
          description: Human-readable error message
          example: "Deal Unique ID is required"
        timestamp:
          type: string
          format: date-time
          description: Timestamp when the error occurred
          example: "2025-11-13T10:30:00"

  securitySchemes: {}

security: []